chore: remove unnecessary files and reorganise directories
