--- src/main/java/uni/usic/domain/entity/tasks/Viewable.java
@@ -1,5 +1,0 @@
-package uni.usic.domain.entity.tasks;
-
-public interface Viewable {
-    void viewable();
-}
