+++ src/main/java/uni/usic/infrastructure/repository/tasks/TaskFileSync.java
@@ -2,7 +2,7 @@ package uni.usic.infrastructure.repository.tasks;
 
 import uni.usic.application.service.tasks.TaskService;
 import uni.usic.domain.entity.tasks.enums.TaskType;
-import uni.usic.domain.entity.tasks.maintasks.Task;
+import uni.usic.domain.entity.tasks.Task;
 import uni.usic.domain.entity.tasks.enums.TaskPriority;
 import uni.usic.domain.entity.tasks.enums.TaskProgress;
