+++ src/main/java/uni/usic/taskmanager/views/tasks/TaskList.java
@@ -15,7 +15,7 @@ import javafx.scene.text.FontWeight;
 import javafx.stage.Stage;
 import uni.usic.application.service.tasks.TaskManager;
 import uni.usic.application.service.tasks.TaskService;
-import uni.usic.domain.entity.tasks.maintasks.Task;
+import uni.usic.domain.entity.tasks.Task;
 import uni.usic.infrastructure.repository.tasks.TaskFileRepository;
 import uni.usic.taskmanager.views.account.ProfileView;
 import uni.usic.taskmanager.views.common.MainMenuBar;
