+++ src/main/java/uni/usic/test/tasks/TestTask.java
@@ -1,7 +1,7 @@
 package uni.usic.test.tasks;
 
 import uni.usic.domain.entity.tasks.enums.TaskType;
-import uni.usic.domain.entity.tasks.maintasks.Task;
+import uni.usic.domain.entity.tasks.Task;
 import uni.usic.domain.entity.tasks.enums.TaskPriority;
 import uni.usic.domain.entity.tasks.enums.TaskProgress;
